pyproject.toml
[tool.poetry]
name = "ing-task"
version = "0.1.0"
description = ""
authors = ["Karim Moradi <kmoradi.ai@gmail.com>"]
readme = "README.md"
#packages = [{include = "ing_task"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
requests = "^2.28.2"
#websocket = "^0.2.1"
#pydbgen = "^1.0.5"
jinja2 = "^3.1.2"
kafka-python = "^2.0.2"
faker = "^17.1.0"
python-dotenv = "^1.0.0"
isort = "^5.12.0"
black = "^23.1.0"
flake8 = "^6.0.0"
pre-commit = "^3.1.0"
pytest = "^7.2.1"
httpx = "^0.23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
